clippy on linux

+++ crates/nu-command/src/dataframe/values/nu_dataframe/conversion.rs
@@ -307,10 +307,7 @@ pub fn create_column(
                 .skip(from_row)
                 .take(size)
                 .map(|v| match v {
-                    Some(a) => Value::Int {
-                        val: a as i64,
-                        span,
-                    },
+                    Some(a) => Value::Int { val: a, span },
                     None => Value::Nothing { span },
                 })
                 .collect::<Vec<Value>>();
